Time array initialization needs to start at 0
By starting at 1 instead of 0 we leave the first value within the t array with an unknown value. I know you are avoiding a divide by 0 so the 0 initialization can be done outside of the for loop.

I was trying to figure out why I was seeing ringing on the oscilloscope right at the zero crossing. 

Turns out the generated signal had a 300mV difference between the last sample and the 0 sample due to t[0] not being 0;